Mendesain ulang struktur file system pada projek: semua file *.c terletak di src, *.h terletak di include
